Backend/requirements.txt
fastapi
sqlmodel
uvicorn
psycopg2-binary
passlib
bcrypt
python-jose
python-dotenv